requirements.txt
pydantic
pyyaml
python-dotenv

